+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - DMA endpoints prefix
 
+### Removed
+
+- Hydraulic engine endpoints
+
 ## [0.6.0] - 2026-01-29
 
 ### Added